requirements/local.txt
-r base.txt

# Testing
django-test-plus==1.0.16
coverage==4.2
django-coverage-plugin==1.3.1
pytest-django==3.0.0
pytest-sugar==0.7.1
factory_boy==2.7.0

# Developing local tools
Werkzeug==0.11.11
ipdb==0.10.1
django-debug-toolbar==1.6
django-extensions==1.7.4

# Documentation
Sphinx==1.4.8

